refactor(config-ui): error handle and db migrate
