+++ config-ui/src/components/index.ts
@@ -24,4 +24,5 @@ export * from './logo';
 export * from './message';
 export * from './no-data';
 export * from './page-header';
+export * from './tip-layout';
 export * from './tooltip';
