+++ config-ui/src/components/tip-layout/styled.ts
