+++ config-ui/src/routes/index.ts
@@ -19,6 +19,7 @@
 export * from './api-keys';
 export * from './blueprint';
 export * from './connection';
+export * from './db-migrate';
 export * from './error';
 export * from './layout';
 export * from './pipeline';
